refactor: split flow designer view and bump beta version
