+++ packages/open-flow/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oomol-lab/open-flow",
-  "version": "0.1.0-beta.6",
+  "version": "0.1.0-beta.7",
   "description": "Portable contracts and shared browser runtime for Open Flow.",
   "keywords": [
     "ai",
